chore: move form logic to a custom hook
